set role and user for login

+++ src/useCases/users.use.js
@@ -1,11 +1,12 @@
 import { StatusHttp } from "../libs/statusHttp.js";
+import { Company } from "../models/companies.models.js";
 import { User } from "../models/users.model.js";
 import bcrypt from "../libs/bcrypt.js";
 
 async function create(newUser) {
   const { email, password } = newUser;
-  const userFound = await User.findOne({ email });
-  if (userFound) throw new StatusHttp("Ya existe este email", 404);
+  const userFound = await User.findOne({ email }) || await Company.findOne({ email });
+  if (userFound) throw new StatusHttp("Ya existe un usuario con este email", 404);
   const encryptedPassword = await bcrypt.hash(password);
   return User.create({ ...newUser, password: encryptedPassword });
 }
@@ -46,12 +47,15 @@ async function update(idUser, newData, file) {
   }
 }
 
-async function deleteById(idUser) {
+async function deleteById(id, idUser) {
   const userFound = await User.findById(idUser);
 
   if (!userFound) throw new StatusHttp("No existe este user", 404);
-
-  return await User.deleteOne({ _id: idUser });
+  if(userFound._id == idUser){
+    return await User.deleteOne({ _id: idUser });
+  }else{
+    throw new StatusHttp("No puedes eliminar un usuario que no te pertenece");
+  }
 }
 
 async function getById(idUser) {
